Refactor webservice / API
